+++ internal/vehicles/domain/commands/save_vehicle_command_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/kalilventura/vehicle-management/internal/vehicles/domain/commands"
 	"github.com/kalilventura/vehicle-management/internal/vehicles/domain/entities"
+	"github.com/kalilventura/vehicle-management/internal/vehicles/domain/entities/dtos"
 	"github.com/kalilventura/vehicle-management/test/vehicles/domain/builders"
 	"github.com/kalilventura/vehicle-management/test/vehicles/infrastructure/repositories"
 	"github.com/stretchr/testify/assert"
@@ -31,7 +32,14 @@ func TestSaveVehicleCommand(t *testing.T) {
 
 	t.Run("should call OnInternalServerError when there's an error to save the new vehicle", func(t *testing.T) {
 		// given
-		vehicle := builders.NewVehicleBuilder().BuildValid()
+		mileage, _ := dtos.NewMileage(0)
+		specification := entities.Specification{
+			Mileage: mileage,
+		}
+		vehicle := builders.NewVehicleBuilder().
+			WithSpecification(specification).
+			WithCondition(dtos.New).
+			Build()
 		repository := repositories.NewInMemoryVehiclesRepository().WithError(errors.New("save vehicle error"))
 		command := commands.NewSaveVehicleCommand(repository)
 
@@ -41,12 +49,19 @@ func TestSaveVehicleCommand(t *testing.T) {
 				assert.Error(t, err)
 			},
 		}
-		command.Execute(vehicle, listeners)
+		command.Execute(&vehicle, listeners)
 	})
 
 	t.Run("should call OnSuccess when the vehicle was saved successfully", func(t *testing.T) {
 		// given
-		vehicle := builders.NewVehicleBuilder().BuildValid()
+		mileage, _ := dtos.NewMileage(0)
+		specification := entities.Specification{
+			Mileage: mileage,
+		}
+		vehicle := builders.NewVehicleBuilder().
+			WithSpecification(specification).
+			WithCondition(dtos.New).
+			Build()
 		repository := repositories.NewInMemoryVehiclesRepository()
 		command := commands.NewSaveVehicleCommand(repository)
 
@@ -56,6 +71,6 @@ func TestSaveVehicleCommand(t *testing.T) {
 				assert.NotNil(t, vehicle)
 			},
 		}
-		command.Execute(vehicle, listeners)
+		command.Execute(&vehicle, listeners)
 	})
 }