+++ internal/vehicles/domain/entities/update_features.go
@@ -1,0 +1,15 @@
+package entities
+
+type UpdateFeaturesInput struct {
+	HasAirConditioning *bool
+	HasAirbag          *bool
+	HasAbsBrakes       *bool
+	HasPowerSteering   *bool
+	HasPowerWindows    *bool
+	HasPowerLocks      *bool
+	HasMultimedia      *bool
+	HasAlarm           *bool
+	HasTractionControl *bool
+	HasRearCamera      *bool
+	HasParkingSensors  *bool
+}