+++ test/vehicles/domain/builders/vehicle_builder.go
@@ -4,12 +4,29 @@ import (
 	global "github.com/kalilventura/vehicle-management/internal/shared/domain/entities"
 	"github.com/kalilventura/vehicle-management/internal/vehicles/domain/entities"
 	"github.com/kalilventura/vehicle-management/internal/vehicles/domain/entities/dtos"
+	"github.com/kalilventura/vehicle-management/test/shared/domain/builders"
 )
 
-type VehicleBuilder struct{}
+type VehicleBuilder struct {
+	builders.BaseBuilder[entities.Vehicle]
+}
 
 func NewVehicleBuilder() *VehicleBuilder {
 	return &VehicleBuilder{}
+}
+
+func (b *VehicleBuilder) WithSpecification(value entities.Specification) *VehicleBuilder {
+	b.AppendModifier(func(e *entities.Vehicle) {
+		e.Specification = value
+	})
+	return b
+}
+
+func (b *VehicleBuilder) WithCondition(value dtos.Condition) *VehicleBuilder {
+	b.AppendModifier(func(e *entities.Vehicle) {
+		e.Condition = value
+	})
+	return b
 }
 
 func (b *VehicleBuilder) BuildValid() *entities.Vehicle {